+++ rust/crd/src/affinity.rs
