+++ rust/crd/src/lib.rs
@@ -1,4 +1,6 @@
-use affinities::get_affinity;
+pub mod affinity;
+
+use affinity::get_affinity;
 use serde::{Deserialize, Serialize};
 use snafu::{OptionExt, ResultExt, Snafu};
 use stackable_operator::{
@@ -20,8 +22,6 @@ use stackable_operator::{
 };
 use std::{collections::BTreeMap, str::FromStr};
 use strum::{Display, EnumIter, EnumString};
-
-mod affinities;
 
 pub const APP_NAME: &str = "hbase";
